review comments: 1. Added hideAfter option 2. css moved to seperate file 3. name changes
